MM-8: security in methods create delete and edit

+++ app/Livewire/VehicleExpense/Delete.php
@@ -30,6 +30,7 @@ class Delete extends Component
 
     public function destroy(): void
     {
+        $this->authorize('expense_delete');
         $this->msg  = 'Expense Deleted';
         $this->icon = 'icons.success';
         $this->form->destroy();
